Add_separate_function_for_parking_fee_calculation
